# docs/swagger.yaml
definitions:
...
  dto.LoginDto:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.RegisterDto:
    properties:
      email:
        type: string
      first_name:
        type: string
      is_admin:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      user_type:
        type: integer
    required:
    - email
    - first_name
    - is_admin
    - password
    - user_type
    type: object
  models.ApplicationUser:
    properties:
      address:
        type: string
      approved_at:
        type: string
      approved_by:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      email:
        type: string
      extension:
        type: string
      fb_uid:
        type: string
      filename:
        type: string
      filepath:
        type: string
      filepath_thumbnail:
        type: string
      first_name:
        type: string
      gender:
        type: boolean
      id:
        type: string
      is_active:
        type: boolean
      is_admin:
        type: boolean
      is_default:
        type: boolean
      is_email_verified:
        type: boolean
      is_locked:
        type: boolean
      is_phone_verified:
        type: boolean
      is_system_admin:
        type: boolean
      last_name:
        type: string
      organization_id:
        type: string
      owner_id:
        type: string
      phone:
        type: string
      size:
        type: string
      submitted_at:
        type: string
      submitted_by:
        type: string
      tenant_id:
        type: string
      title:
        type: string
      token:
        type: string
      total_point:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      user_type:
        type: integer
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /application_user/getAll:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      tags:
      - Application User
  /application_user/getProfile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      tags:
      - Application User
  /application_user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: application_user
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      tags:
      - Application User
  /application_user/update:
    patch:
      consumes:
      - application/json
      parameters:
      - description: application_user
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/models.ApplicationUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - BearerAuth: []
      tags:
      - Application User
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"